# neo4things-api/openapi-schema.yml
openapi: 3.0.3
info:
  title: ''
  version: 0.0.0
servers:
- url: http://localhost:7000
paths:
...
  /mud/:
    get:
      operationId: mud_list
      description: API endpoint that allows Things to be edited or viewed.
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - mud
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMUDList'
          description: ''
    post:
      operationId: mud_create
      description: API endpoint that allows Things to be edited or viewed.
      tags:
      - mud
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MUD'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MUD'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MUD'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MUD'
          description: ''
  /mud/{mud_url}/:
    get:
      operationId: mud_retrieve
      description: API endpoint that allows Things to be edited or viewed.
      parameters:
      - in: path
        name: mud_url
        schema:
          type: string
        required: true
      tags:
      - mud
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MUD'
          description: ''
    put:
      operationId: mud_update
      description: API endpoint that allows Things to be edited or viewed.
      parameters:
      - in: path
        name: mud_url
        schema:
          type: string
        required: true
      tags:
      - mud
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MUD'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MUD'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MUD'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MUD'
          description: ''
    patch:
      operationId: mud_partial_update
      description: API endpoint that allows Things to be edited or viewed.
      parameters:
      - in: path
        name: mud_url
        schema:
          type: string
        required: true
      tags:
      - mud
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedMUD'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedMUD'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedMUD'
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MUD'
          description: ''
    delete:
      operationId: mud_destroy
      description: API endpoint that allows Things to be edited or viewed.
      parameters:
      - in: path
        name: mud_url
        schema:
          type: string
        required: true
      tags:
      - mud
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '204':
          description: No response body
...
  /users/:
    get:
      operationId: users_list
      description: API endpoint that allows users to be viewed or edited.
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - users
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserList'
          description: ''
    post:
      operationId: users_create
      description: API endpoint that allows users to be viewed or edited.
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
  /users/{id}/:
    get:
      operationId: users_retrieve
      description: API endpoint that allows users to be viewed or edited.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this user.
        required: true
      tags:
      - users
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
    put:
      operationId: users_update
      description: API endpoint that allows users to be viewed or edited.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this user.
        required: true
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
    patch:
      operationId: users_partial_update
      description: API endpoint that allows users to be viewed or edited.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this user.
        required: true
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedUser'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedUser'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedUser'
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
    delete:
      operationId: users_destroy
      description: API endpoint that allows users to be viewed or edited.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this user.
        required: true
      tags:
      - users
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '204':
          description: No response body
components:
  schemas:
...
    MUD:
      type: object
      properties:
        mud_version:
          type: integer
        mud_url:
          type: string
          format: uri
        mud_signature:
          type: string
          maxLength: 128
        cache_validity:
          type: integer
          maximum: 168
        systeminfo:
          type: string
          maxLength: 512
        name:
          type: string
          maxLength: 64
        firmware_rev:
          type: string
          maxLength: 5
        software_rev:
          type: string
          maxLength: 5
        documentation:
          type: string
          format: uri
      required:
      - cache_validity
      - documentation
      - firmware_rev
      - mud_signature
      - mud_url
      - mud_version
      - name
      - software_rev
      - systeminfo
...
    PaginatedMUDList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/MUD'
...
    PaginatedUserList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/User'
...
    PatchedMUD:
      type: object
      properties:
        mud_version:
          type: integer
        mud_url:
          type: string
          format: uri
        mud_signature:
          type: string
          maxLength: 128
        cache_validity:
          type: integer
          maximum: 168
        systeminfo:
          type: string
          maxLength: 512
        name:
          type: string
          maxLength: 64
        firmware_rev:
          type: string
          maxLength: 5
        software_rev:
          type: string
          maxLength: 5
        documentation:
          type: string
          format: uri
...
    PatchedUser:
      type: object
      properties:
        url:
          type: string
          format: uri
          readOnly: true
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        groups:
          type: array
          items:
            type: string
            format: uri
          description: The groups this user belongs to. A user will get all permissions
            granted to each of their groups.
...
    User:
      type: object
      properties:
        url:
          type: string
          format: uri
          readOnly: true
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        groups:
          type: array
          items:
            type: string
            format: uri
          description: The groups this user belongs to. A user will get all permissions
            granted to each of their groups.
      required:
      - url
      - username
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: Session