# src/routes/users_management_controller.rs
#![allow(clippy::needless_pass_by_value)]

use isahc::http::StatusCode;
use paperclip::actix::{api_v2_operation, web, web::Json};
use validator::{HasLen, Validate};

use crate::{auth::AuthToken, db::DbConnection, error, error::Result, services::role_service::permission::Permission};

use crate::{
    models::User,
    routes::dtos::{MgmCreateUserDto, MgmUpdateUserBasicDto, MgmUserDto},
    services::{role_service, user_service},
};
use actix_web::HttpResponse;

pub fn init(cfg: &mut web::ServiceConfig) {
    cfg.route("/", web::get().to(get_all_users));
    cfg.route("/", web::post().to(create_user));
    cfg.route("/{user_id}", web::get().to(get_user_by_id));
    cfg.route("/{user_id}", web::put().to(update_user_by_id));
    cfg.route("/{user_id}", web::delete().to(delete_user_by_id));
}

#[api_v2_operation(summary = "List of all users")]
pub async fn get_all_users(pool: web::Data<DbConnection>, auth: AuthToken) -> Result<Json<Vec<MgmUserDto>>> {
    auth.require_permission(Permission::user__management__list)?;

    let mut all_users: Vec<MgmUserDto> = vec![];

    for user in user_service::get_all(pool.get_ref()).await? {
        debug!(">user> {:?}", user);
        all_users.push(MgmUserDto {
            user_id: user.id,
            username: user.username,
            roles: user.roles,
            permissions: user.permissions,
        });
    }

    Ok(Json(all_users))
}

#[api_v2_operation(summary = "Get a specific user")]
pub async fn get_user_by_id(
    pool: web::Data<DbConnection>,
    auth: AuthToken,
    user_id: web::Path<i64>,
) -> Result<Json<MgmUserDto>> {
    auth.require_permission(Permission::user__management__read)?;

    Ok(Json(get_user_from_service(pool.get_ref(), user_id.into_inner()).await?))
}

#[api_v2_operation(summary = "Create a new user")]
pub async fn create_user(
    pool: web::Data<DbConnection>,
    auth: AuthToken,
    create_user_dto: Json<MgmCreateUserDto>,
) -> Result<Json<MgmUserDto>> {
    auth.require_permission(Permission::user__management__create)?;

    create_user_dto.validate().or_else(|_| {
        error::ResponseError {
            status: StatusCode::BAD_REQUEST,
            message: None,
        }
        .fail()
    })?;

    let user = User::new(create_user_dto.username.clone(), create_user_dto.password.as_str())?;
    let new_user_id = user_service::insert(user, pool.get_ref()).await?;

    if auth.require_permission(Permission::role__assign).ok().is_some() {
        for role_id in create_user_dto.roles_ids.iter() {
            role_service::role_add_to_user(pool.get_ref(), new_user_id, *role_id).await?;
        }
    }

    Ok(Json(get_user_from_service(pool.get_ref(), new_user_id).await?))
}

#[api_v2_operation(summary = "Update a user")]
pub async fn update_user_by_id(
    pool: web::Data<DbConnection>,
    auth: AuthToken,
    user_id: web::Path<i64>,
    update_user_dto: Json<MgmUpdateUserBasicDto>,
) -> Result<HttpResponse> {
    auth.require_permission(Permission::user__management__write)?;

    update_user_dto.validate().or_else(|_| {
        error::ResponseError {
            status: StatusCode::BAD_REQUEST,
            message: None,
        }
        .fail()
    })?;

    let user_db_result = user_service::find_by_username(&update_user_dto.username, pool.get_ref()).await;
    if user_db_result.is_ok() {
        let same_name_user_db = user_db_result.unwrap();
        if same_name_user_db.id != user_id.clone() {
            return Ok(HttpResponse::Conflict().reason("Username already in use!").finish());
        }
    }

    let user_db = user_service::find_by_id(user_id.clone(), pool.get_ref()).await?;
    let mut user = User {
        id: user_id.clone(),
        username: update_user_dto.username.clone(),
        password: user_db.password,
        salt: user_db.salt,
        roles: vec![],
        permissions: vec![],
    };

    user_service::update_username(user_id.clone(), &user, pool.get_ref()).await?;

    if update_user_dto.password.length() > 0 {
        user.password = User::hash_password(&update_user_dto.password, &user.salt)?;
        user_service::update_password(user.id, &user, pool.get_ref()).await?;
    }

    Ok(HttpResponse::NoContent().finish())
}

#[api_v2_operation(summary = "Delete a user")]
pub async fn delete_user_by_id(
    pool: web::Data<DbConnection>,
    auth: AuthToken,
    user_id: web::Path<i64>,
) -> Result<HttpResponse> {
    auth.require_permission(Permission::user__management__delete)?;

    user_service::delete(user_id.into_inner(), pool.get_ref()).await?;

    Ok(HttpResponse::NoContent().finish())
}

async fn get_user_from_service(pool: &DbConnection, user_id: i64) -> Result<MgmUserDto> {
    let user = user_service::find_by_id(user_id, pool).await?;

    Ok(MgmUserDto {
        user_id: user.id,
        username: user.username,
        roles: user.roles,
        permissions: user.permissions,
    })
}